feat: show more button (#58)
* fix: reorder IDs

Logical mistake. Newest project should have the highest ID.

* style: clean up commented out code

* feat: create animated slide-out cards, remove overflow-x-hidden from body to make sticky positioning work, add new projects/experiments

* feat: make custom Tailwind colors usable with opacity

Makes the use of "bg-theme-900/50" possible

* feat: show the show more button conditionally, only if there are more items to be shown

* feat: add arbitrary padding to show more buttons and remove button height

This change prevents the button from taking up space in the document. The bottom padding added to the sticky container acts as a way to give the button height. Not the best practice but it's the most straightforward solution.

* fix: improve card style on mobile via smaller icons and arbitrary translateY calculations

All done using media queries. The queries take xs: (450px) as base.

* chore: copy experiments.component.css into blog.component.css

ViewEncapsulation is set to none on the experiments section component, so the CSS doesn't need duplication. Just in case.

* fix: have normal-card handle site routes and external pages

* chore: shorten c-twig-server description