feat: add portal directive and service
This was necessary to have the elements inside the page break out of the layout: A classic SPA issue.

Without portal, any elements wanting to break from the layout would need to use transforms and absolute positioning to look the part. Page headers for example would need to take up full width of the window, instead of the ~800px space given to the page in the layout.

Absolute positioning leads to issues in the DOM. Page header isn't actually higher up in the DOM, it just looks like it does; elements like sidebars that need to sit next to the page need to have extra arbitrary padding to make space for the page header that may render over it.

The portal fixes this issue by moving elements from being rendered in one component to another - especially useful for pages/routes.