feat: move site-logo and vs-search styling, animations, logic to top-bar.component; add new animation for displaying visible elements over the header
Portal's changes observable is used to get notified when the header element is ready and mounted properly; when it is ready, the header element's viewRef is queried, and its clientHeight is received.

When scrollY is 0, no shadow, background color, blur is displayed; blur is displayed after scrollY > 20; when scrollY > clientHeight, shadow and background color are displayed.