feat: add new loading completion trigger/strategy to fake-loading-bar.service.ts, remove loading animation
Previously, the fake loading bar would "complete" when the router reported being done loading the page. This is still true and is the default behavior.  If a page needs to load additional data before being officially marked "loaded", this allows the page to do that by overriding the router's complete() call.

The loading animation wasn't happy with the portal implementation, and I didn't want to bother fixing it. Animation wasn't that good anyway; FCP was bad.