Redesign for modern layout (#64)
* chore: remove view-transition related CSS

The site hasn't been using browser view transitions for a long time, so it makes sense to delete these.

* chore: remove Inter font

The site uses Inter Tight everywhere; Inter was loaded for no reason.

* feat: use Renderer2 to render markdown in markdown-renderer.component.ts and add HeadingDirectives, migrate CSS from markdown-renderer-html.css

There were two markdown renderers: One that recursively mounts itself to get Angular to add nodes to DOM, and another that generates an HTML string and mounts it via [innerHTML].

The first one was more complicated and would pollute the DOM with way too many <markdown-renderer> Angular containers. There is likely no way to remove these from the DOM. It's a pet peeve.

The second one is simpler in logic as it just outputs an HTML string, but performing string operations, then having the output string be checked by Angular for unsafe content may have been a performance hit.

The first component now uses a structural directive called MarkdownRendererDirective which uses Renderer2 instead. DOM looks cleaner as a result. It also manages HeadingDirective of the appended nodes.

* feat: add portal directive and service

This was necessary to have the elements inside the page break out of the layout: A classic SPA issue.

Without portal, any elements wanting to break from the layout would need to use transforms and absolute positioning to look the part. Page headers for example would need to take up full width of the window, instead of the ~800px space given to the page in the layout.

Absolute positioning leads to issues in the DOM. Page header isn't actually higher up in the DOM, it just looks like it does; elements like sidebars that need to sit next to the page need to have extra arbitrary padding to make space for the page header that may render over it.

The portal fixes this issue by moving elements from being rendered in one component to another - especially useful for pages/routes.

* perf: manage -up and -move eventlisteners locally on components, remove functionality from window-observer.service.ts, rename file to window.service.ts

These event listeners were always running under window-observer.service.ts, a massive performance killer. Could have implemented a solution like starting and stopping listening based on number of observable subscribers (removeEventListener on 0 subscribers, addEventListeners if 1+) - but the additional overhead didn't make sense.

* feat: move site-logo and vs-search styling, animations, logic to top-bar.component; add new animation for displaying visible elements over the header

Portal's changes observable is used to get notified when the header element is ready and mounted properly; when it is ready, the header element's viewRef is queried, and its clientHeight is received.

When scrollY is 0, no shadow, background color, blur is displayed; blur is displayed after scrollY > 20; when scrollY > clientHeight, shadow and background color are displayed.

* chore: remove unused pages

All blog and projects should be displayed in /, these pages would have contributed very little. Could be considered in the future if I have a giant list of ramblings and an even larger list of projects

* feat: change highlight colors in Beige Light theme

The previous colors were too muted and blended into the background. New colors are closer to orange.

* feat: add new loading completion trigger/strategy to fake-loading-bar.service.ts, remove loading animation

Previously, the fake loading bar would "complete" when the router reported being done loading the page. This is still true and is the default behavior.  If a page needs to load additional data before being officially marked "loaded", this allows the page to do that by overriding the router's complete() call.

The loading animation wasn't happy with the portal implementation, and I didn't want to bother fixing it. Animation wasn't that good anyway; FCP was bad.

* feat: use is-heading directive for headings in /, remove sidepanel-toc's weird padding styling

* feat: preload my picture, fix live-photo rendering issues (clipping)

Video would occasionally escape the overflow-hidden on the parent, showing the video's edges. Forced layer creation, added the same rounded edges to the video, and made sure to fill parent element

* feat: redesign / and /blog/*, add sidebar TOC, change page layout from flex to grid, switch markdown renderers, remove markdown-editor component from view-blog-page.component

Uses portal to show page header above the router's outlet. Sidebar TOC is not yet visible for mobile.