feat: rewrite markdown parser for performance and readability (#71)
The previous one was a toy. It was sloppily thrown together and was pretty much impossible to fix anything with, or add new features to. This one is a slightly more readable toy.

New version is also more performant - 45% to 65% perf improvement. It properly wraps text in <p> elements. Uses a stack-ish thing because it just makes sense.