fix: draw background on ripple-bg to no longer show black bg on Chrome Android on page scroll
Canvas transparency is ignored on page scroll on Android Chrome, so the background would turn black after scrolling down a bit. This fixes the issue by drawing the background color on canvas.

I had hoped to avoid doing this because fillRect is slower than clearRect. There is a slight delay between pageload and canvas first frame; if canvas alpha was turned off, the canvas would display a black background. So I have chosen the worst of both worlds: Alpha on, and fillRect called every frame.